pkg/gadgets: Move standard gadgets to own package
"Standard Gadgets" are wrappers around BCC scripts. Let's move them
to its own package to keep only gadgets written in go in pkg/gadgets.
