+++ pkg/gadget-collection/gadgets/trace/exec/gadget.go
@@ -24,8 +24,8 @@ import (
 	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/execsnoop/tracer"
 
 	coretracer "github.com/kinvolk/inspektor-gadget/pkg/gadgets/execsnoop/tracer/core"
-	standardtracer "github.com/kinvolk/inspektor-gadget/pkg/gadgets/execsnoop/tracer/standard"
 	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/execsnoop/types"
+	standardtracer "github.com/kinvolk/inspektor-gadget/pkg/standardgadgets/trace/exec"
 
 	gadgetv1alpha1 "github.com/kinvolk/inspektor-gadget/pkg/apis/gadget/v1alpha1"
 )
