+++ pkg/gadget-collection/gadgets/trace/tcpconnect/gadget.go
@@ -24,8 +24,8 @@ import (
 	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/tcpconnect/tracer"
 
 	coretracer "github.com/kinvolk/inspektor-gadget/pkg/gadgets/tcpconnect/tracer/core"
-	standardtracer "github.com/kinvolk/inspektor-gadget/pkg/gadgets/tcpconnect/tracer/standard"
 	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/tcpconnect/types"
+	standardtracer "github.com/kinvolk/inspektor-gadget/pkg/standardgadgets/trace/tcpconnect"
 
 	gadgetv1alpha1 "github.com/kinvolk/inspektor-gadget/pkg/apis/gadget/v1alpha1"
 )
