+++ pkg/standardgadgets/profile/cpu/tracer.go
